color_tracker_robot/requirements.txt
numpy
opencv-python
picamera
RPi.GPIO2